Add import graph and Tauri command coverage analysis
Introduces new modules for analyzing Tauri command coverage and import graph structure, including Rust and JS files for graph visualization. Refactors HTML report generation to modularize rendering of insights, duplicate exports, cascades, dynamic imports, command coverage, and the import graph, and integrates a self-hosted Cytoscape-based graph viewer.
